Rename package models -> model
