+++ src/test/kotlin/com/amazon/opendistroforelasticsearch/indexstatemanagement/TestHelpers.kt
@@ -16,17 +16,17 @@
 package com.amazon.opendistroforelasticsearch.indexstatemanagement
 
 import com.amazon.opendistroforelasticsearch.indexstatemanagement.elasticapi.string
-import com.amazon.opendistroforelasticsearch.indexstatemanagement.models.ChangePolicy
-import com.amazon.opendistroforelasticsearch.indexstatemanagement.models.Conditions
-import com.amazon.opendistroforelasticsearch.indexstatemanagement.models.ManagedIndexConfig
-import com.amazon.opendistroforelasticsearch.indexstatemanagement.models.Policy
-import com.amazon.opendistroforelasticsearch.indexstatemanagement.models.State
-import com.amazon.opendistroforelasticsearch.indexstatemanagement.models.Transition
-import com.amazon.opendistroforelasticsearch.indexstatemanagement.models.actions.ActionRetry
-import com.amazon.opendistroforelasticsearch.indexstatemanagement.models.actions.ActionTimeout
-import com.amazon.opendistroforelasticsearch.indexstatemanagement.models.actions.DeleteActionConfig
-import com.amazon.opendistroforelasticsearch.indexstatemanagement.models.coordinator.ClusterStateManagedIndexConfig
-import com.amazon.opendistroforelasticsearch.indexstatemanagement.models.coordinator.SweptManagedIndexConfig
+import com.amazon.opendistroforelasticsearch.indexstatemanagement.model.ChangePolicy
+import com.amazon.opendistroforelasticsearch.indexstatemanagement.model.Conditions
+import com.amazon.opendistroforelasticsearch.indexstatemanagement.model.ManagedIndexConfig
+import com.amazon.opendistroforelasticsearch.indexstatemanagement.model.Policy
+import com.amazon.opendistroforelasticsearch.indexstatemanagement.model.State
+import com.amazon.opendistroforelasticsearch.indexstatemanagement.model.Transition
+import com.amazon.opendistroforelasticsearch.indexstatemanagement.model.actions.ActionRetry
+import com.amazon.opendistroforelasticsearch.indexstatemanagement.model.actions.ActionTimeout
+import com.amazon.opendistroforelasticsearch.indexstatemanagement.model.actions.DeleteActionConfig
+import com.amazon.opendistroforelasticsearch.indexstatemanagement.model.coordinator.ClusterStateManagedIndexConfig
+import com.amazon.opendistroforelasticsearch.indexstatemanagement.model.coordinator.SweptManagedIndexConfig
 import com.amazon.opendistroforelasticsearch.jobscheduler.spi.schedule.CronSchedule
 import com.amazon.opendistroforelasticsearch.jobscheduler.spi.schedule.IntervalSchedule
 import com.amazon.opendistroforelasticsearch.jobscheduler.spi.schedule.Schedule
