+++ src/test/kotlin/com/amazon/opendistroforelasticsearch/indexstatemanagement/model/ConditionsTests.kt
@@ -15,7 +15,6 @@
 
 package com.amazon.opendistroforelasticsearch.indexstatemanagement.model
 
-import com.amazon.opendistroforelasticsearch.indexstatemanagement.models.Conditions
 import com.amazon.opendistroforelasticsearch.indexstatemanagement.randomByteSizeValue
 import com.amazon.opendistroforelasticsearch.indexstatemanagement.randomTimeValueObject
 import org.elasticsearch.common.unit.TimeValue
